Move Lagrange polys to `_lagrange.py`
