+++ src/galois/_polys/__init__.py
@@ -6,6 +6,7 @@ from ._conway import *
 from ._factor import *
 from ._functions import *
 from ._irreducible import *
+from ._lagrange import *
 from ._poly import *
 from ._primitive import *
