+++ src/galois/_polys/_dense.py
@@ -520,63 +520,3 @@ class roots_jit(Function):
                 break
 
         return np.array([roots, powers], dtype=nonzero_coeffs.dtype)
-
-
-class lagrange_poly_jit(Function):
-    """
-    Finds the roots of the polynomial f(x).
-    """
-
-    def __call__(self, x: Array, y: Array) -> Array:
-        verify_isinstance(x, Array)
-        verify_isinstance(y, Array)
-        if not type(x) == type(y):  # pylint: disable=unidiomatic-typecheck
-            raise TypeError(f"Arguments 'x' and 'y' must be over the same Galois field, not {type(x)} and {type(y)}.")
-        if not x.ndim == 1:
-            raise ValueError(f"Argument 'x' must be 1-D, not have shape {x.shape}.")
-        if not y.ndim == 1:
-            raise ValueError(f"Argument 'y' must be 1-D, not have shape {y.shape}.")
-        if not x.size == y.size:
-            raise ValueError(f"Arguments 'x' and 'y' must be the same size, not {x.size} and {y.size}.")
-        if not x.size == np.unique(x).size:
-            raise ValueError(f"Argument 'x' must have unique entries, not {x}.")
-        dtype = x.dtype
-
-        if self.field.ufunc_mode != "python-calculate":
-            coeffs = self.jit(x.astype(np.int64), y.astype(np.int64))
-            coeffs = coeffs.astype(dtype)
-        else:
-            coeffs = self.python(x.view(np.ndarray), y.view(np.ndarray))
-        coeffs = self.field._view(coeffs)
-
-        return coeffs
-
-    def set_globals(self):
-        # pylint: disable=global-variable-undefined
-        global NEGATIVE, SUBTRACT, MULTIPLY, RECIPROCAL, POLY_ADD, POLY_FLOORDIV, POLY_MULTIPLY
-        NEGATIVE = self.field._negative.ufunc_call_only
-        SUBTRACT = self.field._subtract.ufunc_call_only
-        MULTIPLY = self.field._multiply.ufunc_call_only
-        RECIPROCAL = self.field._reciprocal.ufunc_call_only
-        POLY_ADD = add_jit(self.field).function
-        POLY_FLOORDIV = floordiv_jit(self.field).function
-        POLY_MULTIPLY = self.field._convolve.function
-
-    _SIGNATURE = numba.types.FunctionType(int64[:](int64[:], int64[:]))
-
-    @staticmethod
-    def implementation(x, y):  # pragma: no cover
-        dtype = x.dtype
-        L = np.array([0], dtype=dtype)  # The Lagrange polynomial L(x) = 0
-        k = x.size  # The number of coordinates
-
-        for j in range(k):
-            lj = np.array([y[j]], dtype=dtype)
-            for m in range(k):
-                if m == j:
-                    continue
-                ljm = MULTIPLY(np.array([1, NEGATIVE(x[m])], dtype=dtype), RECIPROCAL(SUBTRACT(x[j], x[m])))
-                lj = POLY_MULTIPLY(lj, ljm)
-            L = POLY_ADD(L, lj)
-
-        return L
