+++ src/galois/_polys/_functions.py
@@ -5,18 +5,10 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from .._domains import Array
-from .._helper import export, verify_isinstance
 from . import _constructors
-from ._dense import lagrange_poly_jit
 
 if TYPE_CHECKING:
     from ._poly import Poly
-
-
-###############################################################################
-# Divisibility
-###############################################################################
 
 
 def gcd(a: Poly, b: Poly) -> Poly:
@@ -104,69 +96,3 @@ def prod(*args: Poly) -> Poly:
         prod_ *= arg
 
     return prod_
-
-
-###############################################################################
-# Special polynomials
-###############################################################################
-
-
-@export
-def lagrange_poly(x: Array, y: Array) -> Poly:
-    r"""
-    Computes the Lagrange interpolating polynomial :math:`L(x)` such that :math:`L(x_i) = y_i`.
-
-    Arguments:
-        x: An array of :math:`x_i` values for the coordinates :math:`(x_i, y_i)`. Must be 1-D. Must have no
-            duplicate entries.
-        y: An array of :math:`y_i` values for the coordinates :math:`(x_i, y_i)`. Must be 1-D. Must be the same
-            size as :math:`x`.
-
-    Returns:
-        The Lagrange polynomial :math:`L(x)`.
-
-    Notes:
-        The Lagrange interpolating polynomial is defined as
-
-        .. math::
-            L(x) = \sum_{j=0}^{k-1} y_j \ell_j(x)
-
-        .. math::
-            \ell_j(x) = \prod_{\substack{0 \le m < k \\ m \ne j}} \frac{x - x_m}{x_j - x_m} .
-
-        It is the polynomial of minimal degree that satisfies :math:`L(x_i) = y_i`.
-
-    References:
-        - https://en.wikipedia.org/wiki/Lagrange_polynomial
-
-    Examples:
-        Create random :math:`(x, y)` pairs in :math:`\mathrm{GF}(3^2)`.
-
-        .. ipython:: python
-
-            GF = galois.GF(3**2)
-            x = GF.elements; x
-            y = GF.Random(x.size); y
-
-        Find the Lagrange polynomial that interpolates the coordinates.
-
-        .. ipython:: python
-
-            L = galois.lagrange_poly(x, y); L
-
-        Show that the polynomial evaluated at :math:`x` is :math:`y`.
-
-        .. ipython:: python
-
-            np.array_equal(L(x), y)
-
-    Group:
-        polys-interpolating
-    """
-    verify_isinstance(x, Array)
-    verify_isinstance(y, Array)
-
-    coeffs = lagrange_poly_jit(type(x))(x, y)
-    poly = Poly(coeffs)
-
-    return poly
